feat: refactoring structure of the module
