+++ ipv4_packets.py
@@ -1,0 +1,13 @@
+import struct
+
+# Unpack IPv4 Packets Recieved
+def ipv4_Packet(data):
+    version_header_len = data[0]
+    version = version_header_len >> 4
+    header_len = (version_header_len & 15) * 4
+    ttl, proto, src, target = struct.unpack('! 8x B B 2x 4s 4s', data[:20])
+    return version, header_len, ttl, proto, ipv4(src), ipv4(target), data[header_len:]
+
+# Returns Formatted IP Address
+def ipv4(addr):
+    return '.'.join(map(str, addr))
