+++ multiple_protocol_methods.py
@@ -1,0 +1,77 @@
+import struct
+import socket
+from format_packets import *
+
+# Unpack Ethernet Frame
+def ethernet_frame(data):
+    dest_mac, src_mac, proto = struct.unpack('! 6s 6s H', data[:14])
+    return get_mac_addr(dest_mac), get_mac_addr(src_mac), socket.htons(proto), data[14:]
+
+    # Format MAC Address
+def get_mac_addr(bytes_addr):
+    bytes_str = map('{:02x}'.format, bytes_addr)
+    mac_addr = ':'.join(bytes_str).upper()
+    return mac_addr
+
+
+
+def icmp_packet_template_method(data):
+    icmp_type, code, checksum, data = icmp_packet(data)
+    print(TAB_1 + 'ICMP Packet:')
+    print(TAB_2 + 'Type: {}, Code: {}, Checksum: {},'.format(icmp_type, code, checksum))
+    print(TAB_2 + 'ICMP Data:')
+    print(format_output_line(DATA_TAB_3, data))
+
+# Unpacks for any ICMP Packet
+def icmp_packet(data):
+    icmp_type, code, checksum = struct.unpack('! B B H', data[:4])
+    return icmp_type, code, checksum, data[4:]
+
+
+def tcp_template_method(raw_data, data):
+    src_port, dest_port, sequence, acknowledgment, flag_urg, flag_ack, flag_psh, flag_rst, flag_syn, flag_fin = struct.unpack('! H H L L H H H H H H', raw_data[:24])
+    print(TAB_1 + 'TCP Segment:')
+    print(TAB_2 + 'Source Port: {}, Destination Port: {}'.format(src_port, dest_port))
+    print(TAB_2 + 'Sequence: {}, Acknowledgment: {}'.format(sequence, acknowledgment))
+    print(TAB_2 + 'Flags:')
+    print(TAB_3 + 'URG: {}, ACK: {}, PSH: {}'.format(flag_urg, flag_ack, flag_psh))
+    print(TAB_3 + 'RST: {}, SYN: {}, FIN:{}'.format(flag_rst, flag_syn, flag_fin))
+
+    if len(data) > 0:
+        # HTTP
+        if src_port == 80 or dest_port == 80:
+            print(TAB_2 + 'HTTP Data:')
+            try:
+                http = HTTP(data)
+                http_info = str(http.data).split('\n')
+                for line in http_info:
+                    print(DATA_TAB_3 + str(line))
+            except:
+                print(format_output_line(DATA_TAB_3, data))
+        else:
+            print(TAB_2 + 'TCP Data:')
+            print(format_output_line(DATA_TAB_3, data))
+
+def udp_template_method(data):
+    src_port, dest_port, length, data = udp_seg(data)
+    print(TAB_1 + 'UDP Segment:')
+    print(TAB_2 + 'Source Port: {}, Destination Port: {}, Length: {}'.format(src_port, dest_port, length))
+
+# Unpacks for any TCP Packet
+def tcp_seg(data):
+    (src_port, destination_port, sequence, acknowledgement, offset_reserved_flag) = struct.unpack('! H H L L H', data[:14])
+    offset = (offset_reserved_flag >> 12) * 4
+    flag_urg = (offset_reserved_flag & 32) >> 5
+    flag_ack = (offset_reserved_flag & 32) >>4
+    flag_psh = (offset_reserved_flag & 32) >> 3
+    flag_rst = (offset_reserved_flag & 32) >> 2
+    flag_syn = (offset_reserved_flag & 32) >> 1
+    flag_fin = (offset_reserved_flag & 32) >> 1
+
+    return src_port, destination_port, sequence, acknowledgement, flag_urg, flag_ack, flag_psh, flag_rst, flag_syn, flag_fin, data[offset:]
+
+
+# Unpacks for any UDP Packet
+def udp_seg(data):
+    src_port, dest_port, size = struct.unpack('! H H 2x H', data[:8])
+    return src_port, dest_port, size, data[8:]
